Moves cli specific code to cli.js
